requirements.txt
asgiref==3.2.8
Django==3.0.7
django-treebeard==4.3.1
pytz==2020.1
sqlparse==0.3.1
